refactor the controller and the model as Integration-Test
